+++ packages/gcf-utils/src/gcf-utils.ts
@@ -18,10 +18,10 @@ import {v1} from '@google-cloud/secret-manager';
 import * as express from 'express';
 // eslint-disable-next-line node/no-extraneous-import
 import {Octokit} from '@octokit/rest';
-import {buildTriggerInfo} from './metrics-logging';
-import {GCFLogger, initLogger} from './gcf-logger';
+import {buildTriggerInfo} from './logging/trigger-parser';
+import {GCFLogger, initLogger} from './logging/gcf-logger';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const LoggingOctokitPlugin = require('../src/logging-octokit-plugin.js');
+const LoggingOctokitPlugin = require('../src/logging/logging-octokit-plugin.js');
 
 const client = new CloudTasksClient();
