+++ packages/gcf-utils/src/logging/trigger-info-builder.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-import {TriggerType} from './gcf-utils';
+import {TriggerType} from '../gcf-utils';
 
 /**
  * Information on GCF execution trigger
