+++ packages/gcf-utils/test/gcf-logger.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import {logger} from '../src/gcf-utils';
-import {GCFLogger, initLogger} from '../src/gcf-logger';
+import {GCFLogger, initLogger} from '../src/logging/gcf-logger';
 import {describe, beforeEach, it} from 'mocha';
 import {ObjectWritableMock} from 'stream-mock';
 import {validateLogs, LogLine, logLevels} from './test-helpers';
